refactor(model): dedup byte_to_2d/pos_2d_to_byte into model::buffer_position (phase 1b)
The conversion `(buffer, byte_offset) -> Position2D` existed verbatim in
two places: `app/clipboard.rs` (one caller) and `input/actions.rs` (three
callers + an inverse `pos_2d_to_byte`). Both copies are byte-identical
pure functions of `&Buffer`.

Consolidate into a new small module `model::buffer_position` exporting both
conversions as free functions. Neither grows `TextBuffer`'s API — both
callers already have a `&Buffer` in hand, so free functions are the right
shape. The new module lives under `model/` so both `app/` and `input/` can
depend on it without crossing layering.

Part of phase 1 (free pure helpers) of
docs/internal/editor-modules-refactor-plan.md.
